+++ store/store_books.rb
@@ -1,0 +1,9 @@
+def store_books
+  books = @books.each_with_index.map do |book, index|
+    { class: book.class, publisher: book.publisher, cover_state: book.cover_state,
+      label: (book.label ? book.label.id : nil),
+      index: index, id: book.id }
+  end
+  json_books = JSON.generate(books)
+  File.write('books.json', json_books)
+end
